Implement API objects
Moves the majority of the logic out of the `eeprom.ino` file and into an
API object. Retains the old communication mechanism as a `TextAPI` and
implements a new API as a binary protocol.

Additionally, fixed an SRP violation in `FileSystem` where it was
directly communicating with the serial interface to provide usage
statistics