Give project a proper name
Project has been renamed, for too long this was simply called
"eeprom-filesystem" it is much more than that now. It has been renamed
"bluefish" with this hardware based component getting the name
"bluefish-firmware". Bluefish is inspired by the name of the encryption
algorithm to be used on the client side, "Twofish", and a play on Dr.
Seuss' story "One fish two fish red fish blue fish".