Simplify Binary API Read For Callers
Previously a caller was expected to get a list of all file ids and read
each one's filename individually to see if matched the filename they
were looking for. This has been improved to only ask the caller to
encrypt the filename first, send it to the device and the device will
return the appropriate file after finding it by name.